+++ src/Soluble/Japha/Bridge/Driver/Pjb62/PjbProxyClient.php
@@ -187,6 +187,7 @@ class PjbProxyClient
             self::$client->throwExceptionProxyFactory = new Proxy\DefaultThrowExceptionProxyFactory(self::$client);
 
             $this->bootstrap();
+
         }
     }